COMDate-class {RDCOMClient}	R Documentation

Classes for representing COM VARIANT values as numbers

Description

Several types of VARIANTs in COM represent their value as a number which is interpreted appropriately based on the type of the VARIANT. For example, a date is the number of days since January 1, 1900, i.e. Midnight Jan 1, 1900 is 2. Similarly, a date-time value contains the hour, minute and second information in the decimal part of the number. In order to be able to associate the VARIANT type with the value when converting the VARIANT to R, we use these classes which provide the numeric value but also class information corresponding to the VARIANT.

Objects from the Class

Objects can be created by calls of the form new("COMDate", ...), etc. They are currently created implicitly in the C code that performs the conversion from VARIANT objects to R. In the future, we will provide facilities for the reverse direction.

Slots

.Data:

Object of class "numeric" this is the numeric vector in which the value is stored.

Extends

Class "numeric", from data part. Class "vector", by class "numeric".

Methods

No methods defined with class "COMDate" in the signature.

Author(s)

Duncan Temple Lang (duncan@r-project.org)

References

http://www.omegahat.net/RDCOMClient http://www.omegahat.net/RDCOMServer http://www.omegahat.net/SWinTypeLibs http://www.omegahat.net/SWinRegistry

See Also

.COM

[Package RDCOMClient version 0.93-0 Index]