.COMInit {RDCOMClient}	R Documentation

Activate and de-activate COM facilities in R

Description

This function allows one to turn on and off the COM facilities in an R session. Generally, one calls this at the start of the session and does not turn it off.

Usage

.COMInit(status = TRUE)

Arguments

status
a logical value with TRUE meaning to activate the COM mechanism and FALSE meaning to turn it off.

Details

This merely calls CoInitialize or CoUninitialize.

Value

NULL.

Author(s)

Duncan Temple Lang (duncan@r-project.org)

References

http://www.omegahat.net/RDCOMClient http://www.omegahat.net/RDCOMServer http://www.omegahat.net/SWinTypeLibs http://www.omegahat.net/SWinRegistry

See Also

COMCreate .COM

Examples

 .COMInit()
 .COMInit(TRUE)

 .COMInit(FALSE)

[Package RDCOMClient version 0.93-0 Index]