COMList {RDCOMClient}	R Documentation

Create an instance of COMList class

Description

This is a function that creates an instance of a COMList-class class or some derived class of this. The COMList-class class allows certain types of COM objects to be treated as an ordered container with a length, much like the R list concept.

Usage

COMList(obj, class = "COMList")

Arguments

obj
the COMIDispatch-class object which is to be treated as a class.

class
the name of the class which is to be instantiated. This argument allows one to use the name of a derived class such as ExcelCOMList or COMTypedList.

Value

An object of the specified class.

Author(s)

Duncan Temple Lang (duncan@r-project.org)

References

http://www.omegahat.net/RDCOMClient http://www.omegahat.net/RDCOMServer http://www.omegahat.net/SWinTypeLibs http://www.omegahat.net/SWinRegistry

See Also

.COM

[Package RDCOMClient version 0.93-0 Index]