CompiledCOMIDispatch-class {RDCOMClient}	R Documentation

DCOM Interface object with knowledge of its methods and their types.

Description

This class is used as a simple extension of COMIDispatch-class to illustrate that it has information about its methods and properties and can make use of this information in coercing arguments and return types apporpriately. This class is used by "compiled" or generated S code that is derived from reading type libraries via the SWinTypeLibs package.

The key things we use this class for are to provide special methods for the $ and [[ operators which can take advantage of the pre-computed information. The class is "never" used directly but rather is extended to created derived classes for particular DCOM interfaces.

Objects from the Class

Objects of this class are not created directly, but typically are implicitly created by calling methods in the compiled/generated code created from the type library processing.

Slots

ref:

Object of class "externalptr", inherited from IUnknown-class

Extends

Class "COMIDispatch", directly. Class "IUnknown", by class "COMIDispatch".

Methods

\$

signature(x = "CompiledCOMIDispatch"): access a property or call a method.

\$<-

signature(x = "CompiledCOMIDispatch", name = "character"): ...

[[

signature(x = "CompiledCOMIDispatch", i = "character"): ...

Author(s)

Duncan Temple Lang (duncan@r-project.org)

References

http://www.omegahat.net/RDCOMClient http://www.omegahat.net/RDCOMServer http://www.omegahat.net/SWinTypeLibs http://www.omegahat.net/SWinRegistry

See Also

COMIDispatch-class generateInterface writeCode

[Package RDCOMClient version 0.93-0 Index]