EnumValue {RDCOMClient}	R Documentation

Create instance of enumeration value class in R

Description

This generic function is used to create an instance of a value from an enumeration constant definition type. See EnumValue-class for more information about these named symbolic constants.

Usage

EnumValue(id, value, obj = new("EnumValue"))

Arguments

id
a string giving the symbolic name of the value.

value
the numeric value for this EnumValue instance.

obj
an instance of the class to be created. This allows the caller to reuse this function with for different derived classes of EnumValue-class. One creates the new object via its default prototype and passes this to EnumValue.

Value

The value of the obj argument with is .Data slot filled in with a named integer vector of length 1.

Author(s)

Duncan Temple Lang (duncan@r-project.org)

See Also

EnumValue-class setAs

[Package RDCOMClient version 0.93-0 Index]