DispatchMethods {RDCOMClient}	R Documentation

S-level constants

Description

Constants used to communicate between S and C code to indicate how a COM call should be invoked. This can be used as the value of the .dispatch argument in the .COM function to control whether to treat the call as a property accessor, method call or both.

Usage

data(DispatchMethods)

References

http://www.omegahat.net/RDCOMClient http://www.omegahat.net/RDCOMServer http://www.omegahat.net/SWinTypeLibs http://www.omegahat.net/SWinRegistry

[Package RDCOMClient version 0.93-0 Index]