SCOMErrorInfo-class {RDCOMClient}	R Documentation

S version of COM Error structure

Description

This is a class that we use to represent an error that was generated in a COM call. If an object of this class is returned, then we know that a call to .COM failed. An object of this class contains information about the error status, a description of the error and a string giving the source of the error (e.g. the name of the COM server).

The C++ code to use this is in place but is currently not activated.

Objects from the Class

Created in the C++ code when a COM call fails.

Slots

status:

Object of class "numeric" the error status for the error. This can be used with system codes to find the symbolic name. See .COMErrors in the package.

source:

Object of class "character" a string giving the source of the error, e.g. the application name.

description:

Object of class "character" the human-readable description of the error.

Methods

No methods defined with class "SCOMErrorInfo" in the signature.

Author(s)

Duncan Temple Lang (duncan@r-project.org)

References

http://www.omegahat.net/RDCOMClient http://www.omegahat.net/RDCOMServer http://www.omegahat.net/SWinTypeLibs

See Also

.COM

[Package RDCOMClient version 0.93-0 Index]