VARIANT-class {RDCOMClient}	R Documentation

Reference to a C-level VARIANT object

Description

The VARIANT class provides a way to represent a C-level VARIANT object in S and assign it to S variable names. The other classes are trivial extensions that provide type information about a variant so that we can dispatch methods on such an object.

Objects from the Class

Currently created in C code when returning values from COM calls. In the future, one will be able to create variant objects directly from R and use them as arguments in .COM calls. Also we will provide methods for accessing the values of the variants, and converting them to different S values.

Slots

ref:

Object of class "externalptr" reference to the C-level address of the variant data structure.

kind:

Object of class "integer" the kind of the variant, e.g. I8, I4, R8, BSTR, DISPATCH, etc. represented as a named integer which has the value from the C-level enumeration.

Methods

No methods defined with class "VARIANT" in the signature.

Author(s)

Duncan Temple Lang (duncan@r-project.org)

References

http://www.omegahat.net/RDCOMClient http://www.omegahat.net/RDCOMServer http://www.omegahat.net/SWinTypeLibs

See Also

.COM

[Package RDCOMClient version 0.93-0 Index]