createCOMReference {RDCOMClient}	R Documentation

Creates S object for COM reference

Description

This is the function that is called anytime a COM object is being created or returned from a C language call to S. This function can examine the object and determine what is the best representation in S for it. It can use the generic IUnknown or COMIDispatch classes to simply represent the pointer value. Alternatively, it might dynamically generate a new S class and accessor methods for accessing properties and functions for that object using the SWinTypeLibs package. Or it might lookup a previously compiled collection of type information and match the GUID of the object's type to find the associated S class name.

This function is not intended to be called from S, but primarily from C code that has access to the COM references.

Usage

createCOMReference(ref, className)

Arguments

ref
the S object that is an external pointer containing the reference to the COM object.

className
The name of the class that is “suggested” by the caller.

Value

An S object that should almost definitely contain the ref value. The class of the object should probably extend IUnknown.

Author(s)

Duncan Temple Lang (duncan@r-project.org)

References

http://www.omegahat.net/RDCOMClient http://www.omegahat.net/RDCOMServer http://www.omegahat.net/SWinTypeLibs

See Also

COMCreate .COM

[Package RDCOMClient version 0.93-0 Index]