getCLSID {RDCOMClient}	R Documentation

Get the UUID/GUID from the human-readable name of an application.

Description

This function provides a way to find the unique identifier for an application or interface given its human-readable form, e.g. "Excel.Application". This is convenient if you want to attempt to load the type library without having an instance of the application or if you want to use this information for looking in the Windows registry ( - see the SWinRegistry package).

Usage

getCLSID(appName)

Arguments

appName
the human-readable string (i.e. character vector of length 1) whose UUID we want to find.

Details

This is a interface to the C routines CLSIDFromString and CLSIDFromProgID

Value

A string which gives the UUID. If the UUID could not be found, this (currently) returns the UUID of all zeros: "{00000000-0000-0000-0000-000000000000}"

Author(s)

Duncan Temple Lang (duncan@r-project.org)

References

http://www.omegahat.net/RDCOMClient http://www.omegahat.net/RDCOMServer http://www.omegahat.net/SWinTypeLibs http://www.omegahat.net/SWinRegistry

Examples

  getCLSID("Excel.Application")

[Package RDCOMClient version 0.93-0 Index]